Add missing w64.c to fix VS 2015 Update 3 link errors
This is just a temporary fix that will get wiped out by the next ffmpeg
roll but it allows testing of the preview version of Update 3 on the try
bots. A permanent fix can be made if needed depending on whether the
final version of Update 3 resolves the link issue or not.

The ultimate problem is that ffmpeg conditionally references a symbol
that is not linked in - essentially going:

     if (CONFIG_W64_DEMUXER && wav->w64)
         left = find_guid(s->pb, ff_w64_guid_data) - 24;

where CONFIG_W64_DEMUXER is #defined to zero, and ff_w64_guid_data is
defined in w64.c.

BUG=264459

Change-Id: I05a9281afd43ca121cb8f7409d31b91c83df343c